+++ packages/merge-styles/src/IStyleSet.ts
@@ -4,8 +4,9 @@ import { IStyleFunctionOrObject, IStyleFunction } from './IStyleFunction';
 /**
  * @deprecated Use `Exclude` provided by TypeScript instead.
  */
-export type Diff<T extends keyof any, U extends keyof any> = ({ [P in T]: P } &
-  { [P in U]: never } & { [x: string]: never })[T];
+export type Diff<T extends keyof any, U extends keyof any> = ({ [P in T]: P } & { [P in U]: never } & {
+  [x: string]: never;
+})[T];
 
 /**
  * @deprecated Use the version provided by TypeScript instead.