+++ packages/react-charting/src/components/CommonComponents/CartesianChart.types.ts
@@ -519,4 +519,13 @@ export interface IModifiedCartesianChartProps extends ICartesianChartProps {
    * Callback method used when mouse leaves the chart boundary.
    */
   onChartMouseLeave?: () => void;
+
+  /** Callback method to get extra margins for domain */
+  getDomainMargins?: (containerWidth: number) => IMargins;
+
+  /** Padding between each bar/line-point */
+  xAxisInnerPadding?: number;
+
+  /** Padding before first bar/line-point and after last bar/line-point */
+  xAxisOuterPadding?: number;
 }