+++ packages/react-components/react-combobox/src/contexts/useListboxContextValues.ts
@@ -4,15 +4,8 @@ import { ComboboxContext } from './ComboboxContext';
 
 export function useListboxContextValues(state: ListboxState): ListboxContextValues {
   const hasComboboxContext = useHasParentContext(ComboboxContext);
-  const {
-    activeOption,
-    focusVisible,
-    multiselect,
-    registerOption,
-    selectedOptions,
-    selectOption,
-    setActiveOption,
-  } = state;
+  const { activeOption, focusVisible, multiselect, registerOption, selectedOptions, selectOption, setActiveOption } =
+    state;
 
   // get register/unregister functions from parent combobox context
   const comboboxRegisterOption = useContextSelector(ComboboxContext, ctx => ctx.registerOption);