+++ packages/react-components/react-datepicker-compat/src/index.ts
@@ -1,0 +1,76 @@
+export { AnimationDirection, Calendar, calendarClassNames, useCalendarStyles_unstable } from './Calendar';
+export type { CalendarNavigationIcons, CalendarProps, CalendarStyleProps, CalendarStyles, ICalendar } from './Calendar';
+
+export { CalendarDay, calendarDayClassNames, useCalendarDayStyles_unstable } from './CalendarDay';
+export type { CalendarDayProps, CalendarDayStyleProps, CalendarDayStyles, ICalendarDay } from './CalendarDay';
+
+export {
+  CalendarDayGrid,
+  calendarDayGridClassNames,
+  extraCalendarDayGridClassNames,
+  useCalendarDayGridStyles_unstable,
+} from './CalendarDayGrid';
+export type {
+  CalendarDayGridProps,
+  CalendarDayGridStyleProps,
+  CalendarDayGridStyles,
+  DayInfo,
+  ICalendarDayGrid,
+  WeekCorners,
+} from './CalendarDayGrid';
+
+export { CalendarMonth, calendarMonthClassNames, useCalendarMonthStyles_unstable } from './CalendarMonth';
+export type { CalendarMonthProps, CalendarMonthStyleProps, CalendarMonthStyles, ICalendarMonth } from './CalendarMonth';
+
+export { calendarPickerClassNames, useCalendarPickerStyles_unstable } from './CalendarPicker';
+export type { CalendarPickerStyleProps, CalendarPickerStyles } from './CalendarPicker';
+
+export { CalendarYear, useCalendarYearStyles_unstable } from './CalendarYear';
+export type {
+  CalendarYearHeaderProps,
+  CalendarYearProps,
+  CalendarYearRange,
+  CalendarYearRangeToString,
+  CalendarYearStrings,
+  CalendarYearStyleProps,
+  CalendarYearStyles,
+  ICalendarYear,
+} from './CalendarYear';
+
+export {
+  DatePicker,
+  datePickerClassNames,
+  defaultDatePickerStrings,
+  renderDatePicker_unstable,
+  useDatePicker_unstable,
+  useDatePickerStyles_unstable,
+} from './DatePicker';
+export type { DatePickerProps, DatePickerStrings, IDatePicker } from './DatePicker';
+
+export {
+  DAYS_IN_WEEK,
+  DateRangeType,
+  DayOfWeek,
+  FirstWeekOfYear,
+  MonthOfYear,
+  TimeConstants,
+  addDays,
+  addMonths,
+  addWeeks,
+  addYears,
+  compareDatePart,
+  compareDates,
+  getDatePartHashValue,
+  getDateRangeArray,
+  getEndDateOfWeek,
+  getMonthEnd,
+  getMonthStart,
+  getStartDateOfWeek,
+  getWeekNumber,
+  getWeekNumbersInMonth,
+  getYearEnd,
+  getYearStart,
+  isInDateRangeArray,
+  setMonth,
+} from './utils';
+export type { CalendarStrings, DateFormatting, DateGridStrings } from './utils';