+++ packages/react-components/react-menu/src/components/MenuItem/MenuItem.tsx
@@ -4,6 +4,7 @@ import { renderMenuItem_unstable } from './renderMenuItem';
 import { useMenuItemStyles_unstable } from './useMenuItemStyles';
 import type { MenuItemProps } from './MenuItem.types';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
+import { useCustomStyleHooks_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * Define a styled MenuItem, using the `useMenuItem_unstable` and `useMenuItemStyles_unstable` hook.
@@ -12,6 +13,10 @@ export const MenuItem: ForwardRefComponent<MenuItemProps> = React.forwardRef((pr
   const state = useMenuItem_unstable(props, ref);
 
   useMenuItemStyles_unstable(state);
+
+  const { useMenuItemStyles_unstable: useCustomStyles } = useCustomStyleHooks_unstable();
+  useCustomStyles(state);
+
   return renderMenuItem_unstable(state);
 });