+++ packages/react-components/react-radio/src/components/Radio/Radio.tsx
@@ -4,6 +4,7 @@ import type { RadioProps } from './Radio.types';
 import { renderRadio_unstable } from './renderRadio';
 import { useRadio_unstable } from './useRadio';
 import { useRadioStyles_unstable } from './useRadioStyles';
+import { useCustomStyleHooks_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * Radio component is a wrapper for a radio button with a label.
@@ -12,6 +13,10 @@ export const Radio: ForwardRefComponent<RadioProps> = React.forwardRef((props, r
   const state = useRadio_unstable(props, ref);
 
   useRadioStyles_unstable(state);
+
+  const { useRadioStyles_unstable: useCustomStyles } = useCustomStyleHooks_unstable();
+  useCustomStyles(state);
+
   return renderRadio_unstable(state);
 });