+++ packages/react-components/react-table/src/components/Table/Table.tsx
@@ -5,6 +5,7 @@ import { useTableStyles_unstable } from './useTableStyles';
 import type { TableProps } from './Table.types';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
 import { useTableContextValues_unstable } from './useTableContextValues';
+import { useCustomStyleHooks_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * Table component
@@ -13,6 +14,10 @@ export const Table: ForwardRefComponent<TableProps> = React.forwardRef((props, r
   const state = useTable_unstable(props, ref);
 
   useTableStyles_unstable(state);
+
+  const { useTableStyles_unstable: useCustomStyles } = useCustomStyleHooks_unstable();
+  useCustomStyles(state);
+
   return renderTable_unstable(state, useTableContextValues_unstable(state));
 });