+++ packages/react-examples/src/react-experiments/TilesList/TilesListPage.tsx
@@ -9,14 +9,17 @@ import {
 } from '@fluentui/react-docsite-components';
 
 import { TilesListBasicExample } from './TilesList.Basic.Example';
-const TilesListBasicExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-experiments/TilesList/TilesList.Basic.Example.tsx') as string;
+const TilesListBasicExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-experiments/TilesList/TilesList.Basic.Example.tsx') as string;
 
 import { TilesListDocumentExample } from './TilesList.Document.Example';
-const TilesListDocumentExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-experiments/TilesList/TilesList.Document.Example.tsx') as string;
+const TilesListDocumentExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-experiments/TilesList/TilesList.Document.Example.tsx') as string;
 
 import { TilesListMediaExample } from './TilesList.Media.Example';
 
-const TilesListMediaExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-experiments/TilesList/TilesList.Media.Example.tsx') as string;
+const TilesListMediaExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react-experiments/TilesList/TilesList.Media.Example.tsx') as string;
 
 export interface ITilesListPageState {
   size: 'small' | 'large';