+++ packages/react-examples/src/react/ProgressIndicator/ProgressIndicator.doc.tsx
@@ -4,8 +4,10 @@ import { ProgressIndicatorBasicExample } from './ProgressIndicator.Basic.Example
 import { IDocPageProps } from '@fluentui/react/lib/common/DocPage.types';
 import { ProgressIndicatorIndeterminateExample } from './ProgressIndicator.Indeterminate.Example';
 
-const ProgressIndicatorBasicExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ProgressIndicator/ProgressIndicator.Basic.Example.tsx') as string;
-const ProgressIndicatorIndeterminateExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ProgressIndicator/ProgressIndicator.Indeterminate.Example.tsx') as string;
+const ProgressIndicatorBasicExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ProgressIndicator/ProgressIndicator.Basic.Example.tsx') as string;
+const ProgressIndicatorIndeterminateExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ProgressIndicator/ProgressIndicator.Indeterminate.Example.tsx') as string;
 
 export const ProgressIndicatorPageProps: IDocPageProps = {
   title: 'ProgressIndicator',